refactor: migrated reformatting logic to separate helper file
